Refactor Cluster class
