+++ clients/src/main/java/org/apache/kafka/clients/admin/internals/AdminMetadataManager.java
@@ -17,6 +17,7 @@
 
 package org.apache.kafka.clients.admin.internals;
 
+import org.apache.kafka.clients.DefaultCluster;
 import org.apache.kafka.clients.MetadataUpdater;
 import org.apache.kafka.common.Cluster;
 import org.apache.kafka.common.Node;
@@ -74,7 +75,7 @@ public class AdminMetadataManager {
     /**
      * The current cluster information.
      */
-    private Cluster cluster = Cluster.empty();
+    private Cluster cluster = DefaultCluster.empty();
 
     /**
      * If we got an authorization exception when we last attempted to fetch
@@ -180,7 +181,7 @@ public class AdminMetadataManager {
     public void clearController() {
         if (cluster.controller() != null) {
             log.trace("Clearing cached controller node {}.", cluster.controller());
-            this.cluster = new Cluster(cluster.clusterResource().clusterId(),
+            this.cluster = new DefaultCluster(cluster.clusterResource().clusterId(),
                 cluster.nodes(),
                 Collections.emptySet(),
                 Collections.emptySet(),
