+++ clients/src/main/java/org/apache/kafka/clients/producer/MockProducer.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.clients.producer;
 
+import org.apache.kafka.clients.DefaultCluster;
 import org.apache.kafka.clients.consumer.OffsetAndMetadata;
 import org.apache.kafka.clients.producer.internals.DefaultPartitioner;
 import org.apache.kafka.clients.producer.internals.FutureRecordMetadata;
@@ -110,7 +111,7 @@ public class MockProducer<K, V> implements Producer<K, V> {
     public MockProducer(final boolean autoComplete,
                         final Serializer<K> keySerializer,
                         final Serializer<V> valueSerializer) {
-        this(Cluster.empty(), autoComplete, new DefaultPartitioner(), keySerializer, valueSerializer);
+        this(DefaultCluster.empty(), autoComplete, new DefaultPartitioner(), keySerializer, valueSerializer);
     }
 
     /**
@@ -122,7 +123,7 @@ public class MockProducer<K, V> implements Producer<K, V> {
                         final Partitioner partitioner,
                         final Serializer<K> keySerializer,
                         final Serializer<V> valueSerializer) {
-        this(Cluster.empty(), autoComplete, partitioner, keySerializer, valueSerializer);
+        this(DefaultCluster.empty(), autoComplete, partitioner, keySerializer, valueSerializer);
     }
 
     /**
@@ -131,7 +132,7 @@ public class MockProducer<K, V> implements Producer<K, V> {
      * Equivalent to {@link #MockProducer(Cluster, boolean, Partitioner, Serializer, Serializer)} new MockProducer(Cluster.empty(), false, null, null, null)}
      */
     public MockProducer() {
-        this(Cluster.empty(), false, null, null, null);
+        this(DefaultCluster.empty(), false, null, null, null);
     }
 
     @Override
