+++ clients/src/main/java/org/apache/kafka/common/Cluster.java
@@ -16,267 +16,46 @@
  */
 package org.apache.kafka.common;
 
-import java.net.InetSocketAddress;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * A representation of a subset of the nodes, topics, and partitions in the Kafka cluster.
  */
-public final class Cluster {
+public interface Cluster {
 
-    private final boolean isBootstrapConfigured;
-    private final List<Node> nodes;
-    private final Set<String> unauthorizedTopics;
-    private final Set<String> invalidTopics;
-    private final Set<String> internalTopics;
-    private final Node controller;
-    private final Map<TopicPartition, PartitionInfo> partitionsByTopicPartition;
-    private final Map<Integer, Node> nodesById;
-    private final ClusterResource clusterResource;
+    Cluster withPartitions(Map<TopicPartition, PartitionInfo> partitions);
 
-    /**
-     * Create a new cluster with the given id, nodes and partitions
-     * @param nodes The nodes in the cluster
-     * @param partitions Information about a subset of the topic-partitions this cluster hosts
-     */
-    public Cluster(String clusterId,
-                   Collection<Node> nodes,
-                   Collection<PartitionInfo> partitions,
-                   Set<String> unauthorizedTopics,
-                   Set<String> internalTopics) {
-        this(clusterId, false, nodes, partitions, unauthorizedTopics, Collections.emptySet(), internalTopics, null);
-    }
+    List<Node> nodes();
 
-    /**
-     * Create a new cluster with the given id, nodes and partitions
-     * @param nodes The nodes in the cluster
-     * @param partitions Information about a subset of the topic-partitions this cluster hosts
-     */
-    public Cluster(String clusterId,
-                   Collection<Node> nodes,
-                   Collection<PartitionInfo> partitions,
-                   Set<String> unauthorizedTopics,
-                   Set<String> internalTopics,
-                   Node controller) {
-        this(clusterId, false, nodes, partitions, unauthorizedTopics, Collections.emptySet(), internalTopics, controller);
-    }
+    Node nodeById(int id);
 
-    /**
-     * Create a new cluster with the given id, nodes and partitions
-     * @param nodes The nodes in the cluster
-     * @param partitions Information about a subset of the topic-partitions this cluster hosts
-     */
-    public Cluster(String clusterId,
-                   Collection<Node> nodes,
-                   Collection<PartitionInfo> partitions,
-                   Set<String> unauthorizedTopics,
-                   Set<String> invalidTopics,
-                   Set<String> internalTopics,
-                   Node controller) {
-        this(clusterId, false, nodes, partitions, unauthorizedTopics, invalidTopics, internalTopics, controller);
-    }
+    Node leaderFor(TopicPartition topicPartition);
 
-    private Cluster(String clusterId,
-                    boolean isBootstrapConfigured,
-                    Collection<Node> nodes,
-                    Collection<PartitionInfo> partitions,
-                    Set<String> unauthorizedTopics,
-                    Set<String> invalidTopics,
-                    Set<String> internalTopics,
-                    Node controller) {
-        this.isBootstrapConfigured = isBootstrapConfigured;
-        this.clusterResource = new ClusterResource(clusterId);
-        // make a randomized, unmodifiable copy of the nodes
-        List<Node> copy = new ArrayList<>(nodes);
-        Collections.shuffle(copy);
-        this.nodes = Collections.unmodifiableList(copy);
-        this.nodesById = new HashMap<>();
-        for (Node node : nodes)
-            this.nodesById.put(node.id(), node);
+    PartitionInfo partition(TopicPartition topicPartition);
 
+    List<PartitionInfo> partitionsForTopic(String topic);
 
-        // index the partitions by topic/partition for quick lookup
-        this.partitionsByTopicPartition = new HashMap<>(partitions.size());
-        for (PartitionInfo p : partitions) {
-            this.partitionsByTopicPartition.put(new TopicPartition(p.topic(), p.partition()), p);
-        }
+    int partitionCountForTopic(String topic);
 
-        this.unauthorizedTopics = Collections.unmodifiableSet(unauthorizedTopics);
-        this.invalidTopics = Collections.unmodifiableSet(invalidTopics);
-        this.internalTopics = Collections.unmodifiableSet(internalTopics);
-        this.controller = controller;
-    }
+    List<PartitionInfo> availablePartitionsForTopic(String topic);
 
-    /**
-     * Create an empty cluster instance with no nodes and no topic-partitions.
-     */
-    public static Cluster empty() {
-        return new Cluster(null, new ArrayList<>(0), new ArrayList<>(0), Collections.emptySet(),
-            Collections.emptySet(), null);
-    }
+    List<PartitionInfo> partitionsForNode(int nodeId);
 
-    /**
-     * Create a "bootstrap" cluster using the given list of host/ports
-     * @param addresses The addresses
-     * @return A cluster for these hosts/ports
-     */
-    public static Cluster bootstrap(List<InetSocketAddress> addresses) {
-        List<Node> nodes = new ArrayList<>();
-        int nodeId = -1;
-        for (InetSocketAddress address : addresses)
-            nodes.add(new Node(nodeId--, address.getHostString(), address.getPort()));
-        return new Cluster(null, true, nodes, new ArrayList<>(0),
-            Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), null);
-    }
+    boolean removePartition(TopicPartition topicPartition);
 
-    /**
-     * Return a copy of this cluster combined with `partitions`.
-     */
-    public Cluster withPartitions(Map<TopicPartition, PartitionInfo> partitions) {
-        Map<TopicPartition, PartitionInfo> combinedPartitions = new HashMap<>(this.partitionsByTopicPartition);
-        combinedPartitions.putAll(partitions);
-        return new Cluster(clusterResource.clusterId(), this.nodes, combinedPartitions.values(),
-                new HashSet<>(this.unauthorizedTopics), new HashSet<>(this.invalidTopics),
-                new HashSet<>(this.internalTopics), this.controller);
-    }
+    Set<String> topics();
 
-    /**
-     * @return The known set of nodes
-     */
-    public List<Node> nodes() {
-        return this.nodes;
-    }
+    Set<String> unauthorizedTopics();
 
-    /**
-     * Get the node by the node id (or null if no such node exists)
-     * @param id The id of the node
-     * @return The node, or null if no such node exists
-     */
-    public Node nodeById(int id) {
-        return this.nodesById.get(id);
-    }
+    Set<String> invalidTopics();
 
-    /**
-     * Get the current leader for the given topic-partition
-     * @param topicPartition The topic and partition we want to know the leader for
-     * @return The node that is the leader for this topic-partition, or null if there is currently no leader
-     */
-    public Node leaderFor(TopicPartition topicPartition) {
-        PartitionInfo info = partitionsByTopicPartition.get(topicPartition);
-        if (info == null)
-            return null;
-        else
-            return info.leader();
-    }
+    Set<String> internalTopics();
 
-    /**
-     * Get the metadata for the specified partition
-     * @param topicPartition The topic and partition to fetch info for
-     * @return The metadata about the given topic and partition, or null if none is found
-     */
-    public PartitionInfo partition(TopicPartition topicPartition) {
-        return partitionsByTopicPartition.get(topicPartition);
-    }
+    boolean isBootstrapConfigured();
 
-    /**
-     * Get the list of partitions for this topic
-     * @param topic The topic name
-     * @return A list of partitions
-     */
-    public List<PartitionInfo> partitionsForTopic(String topic) {
-        return partitionsByTopicPartition.values().stream()
-                .filter(info -> info.topic().equals(topic))
-                .collect(Collectors.toList());
-    }
+    ClusterResource clusterResource();
 
-    /**
-     * Get the number of partitions for the given topic
-     * @param topic The topic to get the number of partitions for
-     * @return The number of partitions or null if there is no corresponding metadata
-     */
-    public int partitionCountForTopic(String topic) {
-        return partitionsForTopic(topic).size();
-    }
-
-    /**
-     * Get the list of available partitions for this topic
-     * @param topic The topic name
-     * @return A list of partitions
-     */
-    public List<PartitionInfo> availablePartitionsForTopic(String topic) {
-        return partitionsByTopicPartition.values().stream()
-                .filter(info -> info.topic().equals(topic))
-                .filter(info -> Objects.nonNull(info.leader()))
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Get the list of partitions whose leader is this node
-     * @param nodeId The node id
-     * @return A list of partitions
-     */
-    public List<PartitionInfo> partitionsForNode(int nodeId) {
-        return partitionsByTopicPartition.values().stream()
-                .filter(info -> Objects.nonNull(info.leader()) && info.leader().id() == nodeId)
-                .collect(Collectors.toList());
-    }
-
-    /**
-     * Remove a partition from this cache.
-     * @param topicPartition
-     * @return
-     */
-    public boolean removePartition(TopicPartition topicPartition) {
-        return partitionsByTopicPartition.remove(topicPartition) != null;
-    }
-
-    /**
-     * Get all topics.
-     * @return a set of all topics
-     */
-    public Set<String> topics() {
-        return this.partitionsByTopicPartition.keySet().stream()
-                .map(TopicPartition::topic)
-                .collect(Collectors.toSet());
-    }
-
-    public Set<String> unauthorizedTopics() {
-        return unauthorizedTopics;
-    }
-
-    public Set<String> invalidTopics() {
-        return invalidTopics;
-    }
-
-    public Set<String> internalTopics() {
-        return internalTopics;
-    }
-
-    public boolean isBootstrapConfigured() {
-        return isBootstrapConfigured;
-    }
-
-    public ClusterResource clusterResource() {
-        return clusterResource;
-    }
-
-    public Node controller() {
-        return controller;
-    }
-
-    @Override
-    public String toString() {
-        return "Cluster(id = " + clusterResource.clusterId() + ", nodes = " + this.nodes +
-            ", partitions = " + this.partitionsByTopicPartition.values() + ", controller = " + controller + ")";
-    }
-
+    Node controller();
 }
