+++ connect/runtime/src/test/java/org/apache/kafka/connect/util/TopicAdminTest.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.connect.util;
 
+import org.apache.kafka.clients.DefaultCluster;
 import org.apache.kafka.clients.NodeApiVersions;
 import org.apache.kafka.clients.admin.MockAdminClient;
 import org.apache.kafka.clients.admin.AdminClientUnitTestEnv;
@@ -121,7 +122,7 @@ public class TopicAdminTest {
         for (int i = 0; i < numNodes; ++i) {
             nodes.put(i, new Node(i, "localhost", 8121 + i));
         }
-        Cluster cluster = new Cluster("mockClusterId", nodes.values(),
+        Cluster cluster = new DefaultCluster("mockClusterId", nodes.values(),
                 Collections.<PartitionInfo>emptySet(), Collections.<String>emptySet(),
                 Collections.<String>emptySet(), nodes.get(0));
         return cluster;
