+++ streams/src/test/java/org/apache/kafka/streams/processor/internals/StreamThreadTest.java
@@ -17,6 +17,8 @@
 package org.apache.kafka.streams.processor.internals;
 
 import java.time.Duration;
+
+import org.apache.kafka.clients.DefaultCluster;
 import org.apache.kafka.clients.admin.MockAdminClient;
 import org.apache.kafka.clients.consumer.Consumer;
 import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
@@ -219,7 +221,7 @@ public class StreamThreadTest {
 
     private Cluster createCluster() {
         final Node node = new Node(0, "localhost", 8121);
-        return new Cluster(
+        return new DefaultCluster(
             "mockClusterId",
             singletonList(node),
             Collections.<PartitionInfo>emptySet(),
