+++ streams/src/test/java/org/apache/kafka/streams/tools/StreamsResetterTest.java
@@ -17,6 +17,7 @@
 package org.apache.kafka.streams.tools;
 
 import kafka.tools.StreamsResetter;
+import org.apache.kafka.clients.DefaultCluster;
 import org.apache.kafka.clients.admin.MockAdminClient;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
@@ -247,7 +248,7 @@ public class StreamsResetterTest {
         for (int i = 0; i < numNodes; ++i) {
             nodes.put(i, new Node(i, "localhost", 8121 + i));
         }
-        return new Cluster("mockClusterId", nodes.values(),
+        return new DefaultCluster("mockClusterId", nodes.values(),
             Collections.<PartitionInfo>emptySet(), Collections.<String>emptySet(),
             Collections.<String>emptySet(), nodes.get(0));
     }
